Add a common header for pins and settings + common I2C class for managing I2C busses across drivers

+++ libraries/pico_explorer/pico_explorer.cpp
@@ -95,7 +95,7 @@ namespace pimoroni {
     }
   }
 
-  void PicoExplorer::set_audio_pin(uint8_t pin) {
+  void PicoExplorer::set_audio_pin(uint pin) {
     pwm_config tone_pwm_cfg = pwm_get_default_config();
 
     // calculate the pwm wrap value for this frequency
